template method, observer and states

+++ 2tdsr/digital-business-enablement/01-design-patterns/src/br/com/fiap/chainOfResponsability/desconto/CalculadoraDeDesconto.java
@@ -34,7 +34,7 @@ public class CalculadoraDeDesconto {
 								new SemDesconto()
 				));
 		
-		return cadeiaDeDesconto.calcular(pedido);
+		return (BigDecimal) cadeiaDeDesconto.calcular(pedido);
 	}
 
 }
